move check to accelerators
