+++ pytorch_lightning/accelerators/ddp_accelerator.py
@@ -311,6 +311,9 @@ class DDPAccelerator(Accelerator):
 
         return results
 
+    def should_return_on_invalid_result(self, result):
+        self.ddp_plugin.should_return_on_invalid_result(result, self.trainer, self.sync_tensor)
+
     def on_before_backward_engine_execution(self):
         self.ddp_plugin.on_before_backward_engine_execution(self.trainer)
 
@@ -318,7 +321,7 @@ class DDPAccelerator(Accelerator):
             self, model: LightningModule, device_ids: List[int]
     ) -> DistributedDataParallel:
         model = self.ddp_plugin.configure_ddp(model, device_ids)
-        self.ddp_plugin.configure_ddp_comm_hook(model, self.trainer)
+        self.ddp_plugin.configure_ddp_comm_hook(model, self.trainer, self.is_single_process_single_device)
         return model
 
     def configure_sync_batchnorm(self, model: LightningModule) -> LightningModule:
@@ -374,4 +377,8 @@ class DDPAccelerator(Accelerator):
 
     @property
     def require_distributed_sampler(self):
+        return True
+
+    @property
+    def is_single_process_single_device(self):
         return True
