+++ pytorch_lightning/accelerators/ddp_cpu_spawn_accelerator.py
@@ -238,10 +238,14 @@ class DDPCPUSpawnAccelerator(Accelerator):
             mp_queue.put(best_model_path)
             mp_queue.put(results)
 
+    def on_before_backward_engine_execution(self):
+        self.ddp_plugin.on_before_backward_engine_execution(self.trainer)
+
     def configure_ddp(
             self, model: LightningModule, device_ids: List[int]
     ) -> DistributedDataParallel:
         model = self.ddp_plugin.configure_ddp(model, device_ids)
+        self.ddp_plugin.configure_ddp_comm_hook(model, self.trainer, self.is_single_process_single_device)
         return model
 
     def configure_sync_batchnorm(self, model: LightningModule) -> LightningModule:
@@ -297,3 +301,8 @@ class DDPCPUSpawnAccelerator(Accelerator):
     @property
     def require_distributed_sampler(self):
         return True
+
+    @property
+    def is_single_process_single_device(self):
+        return True
+
