+++ pytorch_lightning/accelerators/ddp_hpc_accelerator.py
@@ -202,10 +202,17 @@ class DDPHPCAccelerator(Accelerator):
 
         return results
 
+    def should_return_on_invalid_result(self, result):
+        self.ddp_plugin.should_return_on_invalid_result(result, self.trainer, self.sync_tensor)
+
+    def on_before_backward_engine_execution(self):
+        self.ddp_plugin.on_before_backward_engine_execution(self.trainer)
+
     def configure_ddp(
             self, model: LightningModule, device_ids: List[int]
     ) -> DistributedDataParallel:
         model = self.ddp_plugin.configure_ddp(model, device_ids)
+        self.ddp_plugin.configure_ddp_comm_hook(model, self.trainer, self.is_single_process_single_device)
         return model
 
     def configure_sync_batchnorm(self, model: LightningModule) -> LightningModule:
@@ -262,3 +269,8 @@ class DDPHPCAccelerator(Accelerator):
     @property
     def require_distributed_sampler(self):
         return True
+
+    @property
+    def is_single_process_single_device(self):
+        return True
+
