+++ pytorch_lightning/accelerators/ddp_spawn_accelerator.py
@@ -270,6 +270,9 @@ class DDPSpawnAccelerator(Accelerator):
                 atomic_save(model.state_dict(), last_path)
             mp_queue.put(last_path)
 
+    def should_return_on_invalid_result(self, result):
+        self.ddp_plugin.should_return_on_invalid_result(result, self.trainer, self.sync_tensor)
+
     def on_before_backward_engine_execution(self):
         self.ddp_plugin.on_before_backward_engine_execution(self.trainer)
 
@@ -277,6 +280,7 @@ class DDPSpawnAccelerator(Accelerator):
             self, model: LightningModule, device_ids: List[int]
     ) -> DistributedDataParallel:
         model = self.ddp_plugin.configure_ddp(model, device_ids)
+        self.ddp_plugin.configure_ddp_comm_hook(model, self.trainer, self.is_single_process_single_device)
         return model
 
     def configure_sync_batchnorm(self, model: LightningModule) -> LightningModule:
@@ -333,3 +337,8 @@ class DDPSpawnAccelerator(Accelerator):
     @property
     def require_distributed_sampler(self):
         return True
+
+    @property
+    def is_single_process_single_device(self):
+        return True
+
